♻️ refactor : text 코드 분리  #30
