+++ src/components/editor/text/text-canvas-stage.tsx
@@ -1,0 +1,172 @@
+'use client';
+import { useEditorStore } from '@/store/editor.store';
+import React, { useEffect, useRef } from 'react';
+import { Layer, Rect, Stage, Text, Transformer } from 'react-konva';
+import TextEditContent from './text-edit-content';
+import Konva from 'konva';
+
+interface CanvasStageProps {
+  shapeRefs: React.MutableRefObject<Record<string, Konva.Text>>;
+}
+
+const TextCanvasStage: React.FC<CanvasStageProps> = ({ shapeRefs }) => {
+  const textElements = useEditorStore((state) => state.textElements);
+  const selectedElementId = useEditorStore((state) => state.selectedElementId);
+  const editingElementId = useEditorStore((state) => state.editingElementId);
+  const addText = useEditorStore((state) => state.addText);
+  const updateText = useEditorStore((state) => state.updateText);
+  const setSelectedElementId = useEditorStore(
+    (state) => state.setSelectedElementId
+  );
+  const setEditingElementId = useEditorStore(
+    (state) => state.setEditingElementId
+  );
+
+  // Transformer 컴포넌트에 대한 ref
+  const transformerRef = useRef<Konva.Transformer | null>(null);
+
+  /**
+   * 선택된 요소가 변경될 때 Transformer의 노드를 업데이트
+   */
+  useEffect(() => {
+    const transformer = transformerRef.current;
+    if (!transformer) return;
+    if (selectedElementId) {
+      const selectedNode = shapeRefs.current[selectedElementId];
+      if (selectedNode) {
+        transformer.nodes([selectedNode]);
+        transformer.getLayer()?.batchDraw();
+      }
+    } else {
+      transformer.nodes([]);
+      transformer.getLayer()?.batchDraw();
+    }
+  }, [selectedElementId]);
+
+  /**
+   * 텍스트 변환 종료 시 업데이트
+   *
+   * @param  id - 변환된 요소의 id.
+   * @param  e - 변환 종료 이벤트 객체
+   */
+  const handleTransformEnd = (
+    id: string,
+    e: Konva.KonvaEventObject<Event>
+  ): void => {
+    const node = e.target;
+    const scaleX = node.scaleX();
+    node.scaleX(1);
+    node.scaleY(1);
+
+    updateText(id, {
+      x: node.x(),
+      y: node.y(),
+      width: Math.max(10, node.width() * scaleX),
+      rotation: node.rotation(),
+    });
+  };
+
+  /**
+   * 더블 클릭 시 편집 모드 실행 핸들러
+   * @param id
+   */
+  const handleTextDoubleClick = (id: string) => {
+    setSelectedElementId(id);
+    setEditingElementId(id);
+  };
+
+  // 인라인 편집 완료 후 텍스트 업데이트
+  const handleTextEditSubmit = (newText: string) => {
+    if (!selectedElementId) return;
+    updateText(selectedElementId, { text: newText });
+    setEditingElementId(null);
+  };
+
+  return (
+    <div className='relative flex flex-1 items-center justify-center bg-white'>
+      <Stage
+        width={800}
+        height={600}
+        onMouseDown={(e) => {
+          if (e.target === e.target.getStage()) {
+            setSelectedElementId(null);
+            setEditingElementId(null);
+          }
+        }}
+        className='border-2'
+      >
+        <Layer>
+          <Rect
+            x={0}
+            y={0}
+            fill='#f9f9f9'
+            onMouseDown={() => {
+              setSelectedElementId(null);
+              setEditingElementId(null);
+            }}
+          />
+          {textElements.map((el) =>
+            el.type === 'text' ? (
+              <Text
+                key={el.id}
+                text={el.text}
+                x={el.x}
+                y={el.y}
+                rotation={el.rotation}
+                fontSize={el.fontSize}
+                fill={el.fill}
+                fontFamily={el.fontFamily}
+                width={el.width}
+                draggable
+                fontStyle={
+                  // 예: bold + italic => 'bold italic'
+                  // 둘 다 false면 'normal'
+                  (el.isBold ? 'bold ' : '') + (el.isItalic ? 'italic' : '') ||
+                  'normal'
+                }
+                textDecoration={[
+                  el.isUnderline ? 'underline' : '',
+                  el.isStrike ? 'line-through' : '',
+                ]
+                  .join(' ')
+                  .trim()}
+                visible={editingElementId !== el.id}
+                onClick={() => setSelectedElementId(el.id)}
+                onTap={() => setSelectedElementId(el.id)}
+                onDblClick={() => handleTextDoubleClick(el.id)}
+                onDblTap={() => handleTextDoubleClick(el.id)}
+                onTransformEnd={(e) => handleTransformEnd(el.id, e)}
+                ref={(node) => {
+                  if (node) {
+                    shapeRefs.current[el.id] = node;
+                  }
+                }}
+              />
+            ) : null
+          )}
+          <Transformer
+            ref={transformerRef}
+            enabledAnchors={['middle-left', 'middle-right']}
+            rotationSnaps={[0, 90, 180, 270]}
+            rotationSnapTolerance={30}
+            rotateEnabled={true}
+          />
+
+          {editingElementId && shapeRefs.current[editingElementId] && (
+            <TextEditContent
+              textNode={shapeRefs.current[editingElementId]}
+              initialText={
+                textElements.find((el) => el.id === editingElementId)?.text ||
+                ''
+              }
+              onChange={handleTextEditSubmit}
+              onClose={() => setEditingElementId(null)}
+            />
+          )}
+        </Layer>
+      </Stage>
+    </div>
+  );
+};
+
+export default TextCanvasStage;
